// src/app/store/recipes/recipes.action-types.ts
export enum ActionTypes {
  GET_TRENDING_RECIPES = 'GET_TRENDING_RECIPES',
  GET_TRENDING_RECIPES_SUCCEED = 'GET_TRENDING_RECIPES_SUCCEED',
  GET_TRENDING_RECIPES_FAILED = 'GET_TRENDING_RECIPES_FAILED',
  GET_RECIPE_BY_NAME = 'GET_RECIPE_BY_NAME',
  GET_RECIPE_BY_NAME_SUCCEED = 'GET_RECIPE_BY_NAME_SUCCEED',
  GET_RECIPE_BY_NAME_FAILED = 'GET_RECIPE_BY_NAME_FAILED',
  GET_RECIPE_BY_ID = 'GET_RECIPE_BY_ID',
  GET_RECIPE_BY_ID_SUCCEED = 'GET_RECIPE_BY_ID_SUCCEED',
  GET_RECIPE_BY_ID_FAILED = 'GET_RECIPE_BY_ID_FAILED',
}